requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
Pillow==10.0.0
rembg==2.0.50
